+++ src/components/common/index.ts
@@ -1,0 +1,6 @@
+export { default as AnswerInput } from "./AnswerInput";
+export { default as Button } from "./Button";
+export { default as Counter } from "./Counter";
+export { default as ModelAnswer } from "./ModelAnswer";
+export { default as Question } from "./Question";
+export { default as Timer } from "./Timer";
