+++ src/pages/InterView.tsx
@@ -1,12 +1,8 @@
-import Timer from "../components/common/Timer";
-import Button from "../components/common/Button";
-import Counter from "../components/common/Counter";
-import Question from "../components/common/Question";
-import AnswerInput from "../components/common/AnswerInput";
 import { getQuestion, postUserAnswer } from "../utils/api";
 import { useMutation, useQuery } from "@tanstack/react-query";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
+import { AnswerInput, Button, Counter, Question, Timer } from "../components/common";
 
 const Interview = () => {
   const questionId = 1; // 현재 MVP에서는 1번 질문만 조회 가능
